feat: dynamic support URLs (WPB-27031)
